Samples/Frameworks/chainlit/SQL-Agent/requirements.txt
# Python General
python-dotenv

# XPANDER
xpander_utils
xpander-sdk

# CHAINLIT
chainlit

# LLMs
openai

# DATA
llama-index-llms-openai
llama-index

# SQL
sqlalchemy
sqlalchemy-bigquery